DownloadFromUpName/requirements.txt
#Command:
#pip install -r .\requirements.txt
python-dotenv==0.19.2
mysql-connector-python==8.0.26